+++ windows/flutter/generated_plugins.cmake
@@ -10,6 +10,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   hotkey_manager_windows
   irondash_engine_context
   media_kit_libs_windows_audio
+  objectbox_flutter_libs
   permission_handler_windows
   record_windows
   screen_retriever_windows